Source/requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
patsy==0.5.1
plotly==4.14.3
plotly-express==0.4.1
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scipy==1.6.3
six==1.16.0
statsmodels==0.12.2
Werkzeug==2.0.1
gunicorn==19.7.1
cassandra-driver==3.28.0
beautifulsoup4
selenium
request
html2text